Save version identifiers with results
This is implemented as saving the commit hash of the master branch of each
submission.

Both hashes for studentA and studentB are just stored as individual
columns, however it may make more sense to save these in a separate table
in the future and map these "submissions" to the results (the current
submissions table).